+++ tests/Mocks/GetPaymentRequestsResponseMock.php
@@ -45,7 +45,7 @@ class GetPaymentRequestsResponseMock
                 '*.abnamro.com/v2/tikkie/paymentrequests/*' => Http::response(
                     [
                         'paymentRequestToken' => $paymentRequestToken,
-                        'amountInCents' => $amount * 100,
+                        'amountInCents' => (int) ($amount * 100),
                         'referenceId' => $referenceId,
                         'description' => $description,
                         'url' => $url,
